Update Styling to match pivotal
CSS has been changed to reflect Pivotal colours (loosely). People are
already familiar with this scheme so it makes it easier to identify what
the status is. I also added story ID for discussion purposes and links to
the full details on pivotal, all the payload for discussion is there.