Improve code structure for readability and understandability
I removed the need for multiple views that only differed by message and status used
for CSS for the view.

I also removed repeated code in the async parallel calls, it was the same apart from the status.
I also renamed a lot of the methods because they were not representative of what they were doing.